+++ src/Http/Requests/UpdateMasterProductVariant.php
@@ -27,6 +27,7 @@ class UpdateMasterProductVariant extends FormRequest implements UpdateMasterProd
             'price' => 'nullable|numeric',
             'original_price' => 'sometimes|nullable|numeric',
             'stock' => 'nullable|numeric',
+            'backorder' => 'nullable|numeric|min:0',
             'excerpt' => 'sometimes|nullable|max:8192',
             'description' => 'sometimes|nullable|max:32768',
             'images' => 'nullable',