+++ src/resources/views/master-product-variant/edit.blade.php
@@ -48,3 +48,8 @@
 
 </div>
 @stop
+
+@push('onload-scripts')
+    const tooltipTriggerList = document.querySelectorAll('[data-bs-toggle="tooltip"]')
+    const tooltipList = [...tooltipTriggerList].map(tooltipTriggerEl => new bootstrap.Tooltip(tooltipTriggerEl))
+@endpush()