+++ src/resources/views/product/create.blade.php
@@ -30,3 +30,8 @@
     {!! Form::close() !!}
 
 @stop
+
+@push('onload-scripts')
+    const tooltipTriggerList = document.querySelectorAll('[data-bs-toggle="tooltip"]')
+    const tooltipList = [...tooltipTriggerList].map(tooltipTriggerEl => new bootstrap.Tooltip(tooltipTriggerEl))
+@endpush()