+++ src/resources/views/product/edit.blade.php
@@ -35,3 +35,8 @@
 
 </div>
 @stop
+
+@push('onload-scripts')
+    const tooltipTriggerList = document.querySelectorAll('[data-bs-toggle="tooltip"]')
+    const tooltipList = [...tooltipTriggerList].map(tooltipTriggerEl => new bootstrap.Tooltip(tooltipTriggerEl))
+@endpush()